+++ storybook/stories/components/LogsProgress.stories.js
@@ -22,6 +22,13 @@ import LogsProgress from "../../../views/templates/components/logs-summary.html.
 export default {
   component: LogsProgress,
   title: "Components/Logs progress",
+};
+
+export const UpdateLogsProgress = {
+  args: {
+    progressStatus: "Update in progress",
+    progressPercentage: 25,
+  },
   play: async ({ args }) => {
     const textSlots = document.querySelectorAll("[data-slot-component='text']");
     textSlots.forEach((slot) => {
@@ -30,16 +37,15 @@ export default {
   },
 };
 
-export const Default = {
-  args: {
-    progressStatus: "Backup files in progress XX files left",
-    progressPercentage: 25,
-  },
-};
-
 export const RestoreLogsProgress = {
   args: {
     progressStatus: "Restoration in progress",
     progressPercentage: 45,
+  },
+  play: async ({ args }) => {
+    const textSlots = document.querySelectorAll("[data-slot-component='text']");
+    textSlots.forEach((slot) => {
+      slot.textContent = args.progressStatus;
+    });
   },
 };
